name of the Irish language
The names (both English and native) of `ga-IE` here are nonsensical. I propose to change `ga-IE` and to make it the same as `ga` because `IE` is the default country for `ga`.

For more information on the Irish language and its name(s), see for example here: http://www.lexiconista.com/pdf/LocalizationWithIrish.pdf

Thank you.